# auditflow-be/src/main/resources/openapi/openapi-audit-v1.yaml
openapi: 3.0.0
info:
  title: Labs64 AuditFlow API
  version: 1.0.0
  description: API for publishing audit events to a message broker for asynchronous processing through configurable pipelines.
  contact:
    name: Labs64
    url: https://labs64.io
    email: info@labs64.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
  - url: /api/v1
    description: Default API base path
tags:
  - name: AuditEvent
    description: Operations to manage and publish audit events.
paths:
  /audit/publish:
    post:
      summary: Publish an audit event
      operationId: publishEvent
      tags:
        - AuditEvent
      description: Publishes an audit event to a message broker for asynchronous processing through configured pipelines.
      security:
        - bearerAuth: []
      requestBody:
        description: Audit event payload to be published.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditEvent'
            example:
              eventId: "123e4567-e89b-12d3-a456-426614174000"
              eventType: "user.login"
              sourceSystem: "auth-service"
              tenantId: "T123456"
              geolocation:
                lat: 48.1264019
                lon: 11.5407647
                countryCode: "DE"
                country: "Germany"
                city: "Munich"
              extra:
                userId: "user123"
                sessionId: "sess456"
                action_name: "login"
                action_status: "SUCCESS"
                action_message: "User logged in successfully"
      responses:
        '200':
          description: Audit event published successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Audit event published successfully"
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "VALIDATION_ERROR"
                message: "eventType must not be null"
                timestamp: "2025-11-30T10:15:30Z"
        '500':
          description: Failed to publish audit event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "INTERNAL_ERROR"
                message: "An internal error occurred while processing your request"
                timestamp: "2025-11-30T10:15:30Z"
components:
  schemas:
    AuditEvent:
      type: object
      description: Represents an audit event.
      required:
        - eventType
        - sourceSystem
      properties:
        timestamp:
          type: string
          format: date-time
          description: The precise time the audit event occurred, in ISO 8601 format.
          example: "YYYY-MM-DDThh:mm:ssZ"
          readOnly: true
        eventId:
          type: string
          format: uuid
          description: A globally unique identifier for this specific audit event.
          example: "fedcba98-7654-3210-fedc-ba9876543210"
        eventType:
          type: string
          description: A high-level classification of the event type (e.g., 'audit.action.performed').
          example: "api.call"
        sourceSystem:
          type: string
          description: Identifies the system or service that generated the audit event.
          example: "netlicensing/core"
        tenantId:
          type: string
          description: Identifier for the tenant or organization associated with the event.
          example: "V12345678"
        geolocation:
          $ref: '#/components/schemas/Geolocation'
        extra:
          $ref: '#/components/schemas/Extra'
    Geolocation:
      type: object
      required:
        - lat
        - lon
      properties:
        lat:
          type: number
          format: double
          description: The latitude coordinate (e.g., 48.1264019).
          example: 48.1264019
        lon:
          type: number
          format: double
          description: The longitude coordinate (e.g., 11.5407647).
          example: 11.5407647
        countryCode:
          type: string
          description: The ISO 3166-1 alpha-2 country code (e.g., "DE").
          example: "DE"
        country:
          type: string
          description: The full country name (e.g., "Germany").
          example: "Germany"
        region:
          type: string
          description: The region, state, or province name (e.g., "Bavaria").
          example: "Bavaria"
        city:
          type: string
          description: The city name where the event originated (e.g., "Munich").
          example: "Munich"
    Extra:
      type: object
      description: Custom user-defined fields.
      additionalProperties: true
      example:
        userId: "customer123"
        userBU: "BU12345678"
        action_name: "login"
        action_status: "SUCCESS"
        action_message: "User logged in successfully"

    ErrorCode:
      type: string
      description: Unique error code identifying the type of error.
      enum:
        - NOT_FOUND
        - INTERNAL_ERROR
        - VALIDATION_ERROR
      example: "NOT_FOUND"

    ErrorResponse:
      type: object
      description: Standard error response returned by the API.
      required:
        - code
        - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Human-readable message describing the error.
          example: "Status is invalid. Allowed values: SUCCESS, FAILURE, PENDING"
        timestamp:
          type: string
          format: date-time
          description: Time when the error occurred.
          example: "YYYY-MM-DDThh:mm:ssZ"
        traceId:
          type: string
          description: Optional trace ID for correlating logs or debugging.
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token